Replace use of double for float

+++ Potato Engine/Engine/Collider.cpp
@@ -35,7 +35,7 @@ Vector2 Collider::getWorldPosition()
 }
 
 
-double Collider::getWorldRotation()
+float Collider::getWorldRotation()
 {
 	return gameObject()->transform->getWorldRotation();
 }
